Switched to f64

+++ krust/src/collision_handler.rs
@@ -4,14 +4,14 @@ use ncollide3d::bounding_volume::{self, BoundingVolume};
 use ncollide3d::shape::Cuboid;
 pub struct CollisionHandler {
 
-    arm_colliders: Vec<Cuboid<f32>>,
-    world_colliders: Vec<Cuboid<f32>>,
+    arm_colliders: Vec<Cuboid<f64>>,
+    world_colliders: Vec<Cuboid<f64>>,
 
 }
 
 impl CollisionHandler{
 
-    pub fn new(arm: Vec<Cuboid<f32>>, world: Vec<Cuboid<f32>>) -> CollisionHandler {
+    pub fn new(arm: Vec<Cuboid<f64>>, world: Vec<Cuboid<f64>>) -> CollisionHandler {
         CollisionHandler {
             arm_colliders: arm,
             world_colliders: world,
@@ -19,7 +19,7 @@ impl CollisionHandler{
     }
 
     // utility functions here
-    pub fn is_arm_colliding(&self, matrices: &Vec<Matrix4<f32>>) -> bool {
+    pub fn is_arm_colliding(&self, matrices: &Vec<Matrix4<f64>>) -> bool {
         // TODO:
         // Matrix4 to isometry
         // Update transform of colliders
